Backend/validation middlewares refactoring (#65)
* Changed types and schemas.

* Folder structure and exports.

* First stage refactor completed.

* Second stage routes refactoring.

* Api docs modification v1.

+++ server/src/api/workspaces/workspace.service.ts
@@ -3,13 +3,10 @@ import type {
   CreateWorkspaceInput,
   UpdateWorkspaceInput,
   WorkspaceDto,
-} from "@ronmordo/types";
+  WorkspaceMemberDto,
+} from "@ronmordo/contracts";
 import { prisma } from "../../lib/prismaClient.js";
-import {
-  type Workspace,
-  type WorkspaceMember,
-  WorkspaceRole,
-} from "@prisma/client";
+import { type Workspace, type WorkspaceMember } from "@prisma/client";
 import {
   mapWorkspaceDtoToCreateInput,
   mapWorkspaceDtoToUpdateInput,
@@ -17,6 +14,7 @@ import {
 } from "./workspace.mapper.js";
 import { AppError } from "../../utils/appError.js";
 import { mapBoardToDto } from "../boards/board.mapper.js";
+import { mapWorkspaceRoleDto } from "../workspace-members/workspace-members.mapper.js";
 
 const createWorkspace = async (
   workspaceDto: CreateWorkspaceInput,
@@ -28,10 +26,14 @@ const createWorkspace = async (
     data: {
       ...workspaceData,
       creator: { connect: { id: creatorId } },
+      workspaceMembers: {
+        create: {
+          userId: creatorId,
+          role: "Admin",
+        },
+      },
     },
   });
-
-  await addWorkspaceMember(workspace.id, workspace.createdBy, "Admin");
 
   return mapWorkspaceToDto(workspace);
 };
@@ -70,7 +72,7 @@ const getWorkspaceWithMembers = async (
   };
 };
 
-const getWorkspacesByUser = async (userId: string): Promise<Workspace[]> => {
+const getWorkspacesByUser = async (userId: string): Promise<WorkspaceDto[]> => {
   const workspaces = await prisma.workspace.findMany({
     where: {
       workspaceMembers: {
@@ -84,7 +86,7 @@ const getWorkspacesByUser = async (userId: string): Promise<Workspace[]> => {
     },
   });
 
-  return workspaces;
+  return workspaces.map(mapWorkspaceToDto);
 };
 
 const getWorkspacesByCreator = async (
@@ -133,14 +135,14 @@ const getAllWorkspaces = async (): Promise<Workspace[]> => {
 // Workspace Member Management
 const addWorkspaceMember = async (
   workspaceId: string,
-  userId: string,
-  role: WorkspaceRole = WorkspaceRole.Member
+  newMemberId: string,
+  role: WorkspaceMemberDto["role"]
 ): Promise<WorkspaceMember> => {
   const member = await prisma.workspaceMember.create({
     data: {
       workspaceId,
-      userId,
-      role,
+      userId: newMemberId,
+      role: mapWorkspaceRoleDto(role),
     },
   });
   console.log("member from service", member);
@@ -149,44 +151,54 @@ const addWorkspaceMember = async (
 
 const removeWorkspaceMember = async (
   workspaceId: string,
-  userId: string
-): Promise<boolean> => {
-  try {
-    await prisma.workspaceMember.delete({
+  authUserId: string,
+  memberIdToDelete: string
+) => {
+  const authMember = await prisma.workspaceMember.findUnique({
+    where: {
+      workspaceId_userId: {
+        workspaceId,
+        userId: authUserId,
+      },
+    },
+  });
+
+  if (!authMember) {
+    throw new AppError("You are not member of this workspace", 403);
+  }
+
+  if (memberIdToDelete === authUserId || authMember.role === "Admin") {
+    return prisma.workspaceMember.delete({
       where: {
         workspaceId_userId: {
           workspaceId,
-          userId,
+          userId: memberIdToDelete,
         },
       },
     });
-    return true;
-  } catch (error) {
-    return false;
   }
+
+  throw new AppError("You are not authorized to remove this member", 403);
 };
 
 const updateWorkspaceMemberRole = async (
   workspaceId: string,
   userId: string,
-  newRole: WorkspaceRole
+  newRole: WorkspaceMemberDto["role"]
 ): Promise<WorkspaceMember | null> => {
-  try {
-    const member = await prisma.workspaceMember.update({
-      where: {
-        workspaceId_userId: {
-          workspaceId,
-          userId,
-        },
+  const member = await prisma.workspaceMember.update({
+    where: {
+      workspaceId_userId: {
+        workspaceId,
+        userId,
       },
-      data: {
-        role: newRole,
-      },
-    });
-    return member;
-  } catch (error) {
-    return null;
-  }
+    },
+    data: {
+      role: mapWorkspaceRoleDto(newRole),
+    },
+  });
+
+  return member;
 };
 
 const getWorkspaceMembers = async (
